Move `BuiltPath` to its own header/C++ file in libcmd
It is less important, and used less widely, than `DerivedPath`.
